requirements.txt
selene>=2.0.0rc7
pytest
allure-pytest
allure-python-commons~=2.13.5
